delete neovim folder after build

+++ install/install-neovim.sh
@@ -12,10 +12,10 @@ function die() {
 }
 
 currDir=$(basename $(pwd))
-
 if [ "$currDir" != "neovim" ]; then
     git clone --depth 1 --branch ${branch}  https://github.com/neovim/neovim || die "Failed to clone neovim"
-    cd neovim
+    pushd neovim
+    delete=1
 else
     echo "Already in neovim directory"
 fi
@@ -36,5 +36,9 @@ ${SUDO} update-alternatives --install /usr/local/bin/nvim nvim /opt/nvim/bin/nvi
 # ${SUDO} update-alternatives --install /usr/local/bin/vim vim /usr/local/bin/nvim 100
 # ${SUDO} update-alternatives --install /usr/local/bin/vi vi /usr/local/bin/nvim 100
 hash -r  # reload hash table so that new version of nvim is found
-rm -rf neovim
+
+if [[ $delete -eq 1 ]]; then
+    popd
+    rm -rf neovim
+fi
